Break up these classes into separate files
