+++ storage/src/main/scala/uk/ac/wellcome/storage/store/dynamo/DynamoHybridStore.scala
@@ -2,7 +2,8 @@ package uk.ac.wellcome.storage.store.dynamo
 
 import java.util.UUID
 
-import uk.ac.wellcome.storage.{S3ObjectLocation, S3ObjectLocationPrefix, Version}
+import uk.ac.wellcome.storage.Version
+import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 import uk.ac.wellcome.storage.store.s3.S3TypedStore
 import uk.ac.wellcome.storage.store.{HybridIndexedStoreEntry, HybridStore}
