+++ storage/src/main/scala/uk/ac/wellcome/storage/store/dynamo/DynamoHybridStoreWithMaxima.scala
@@ -2,9 +2,10 @@ package uk.ac.wellcome.storage.store.dynamo
 
 import java.util.UUID
 
-import uk.ac.wellcome.storage.{S3ObjectLocation, S3ObjectLocationPrefix, Version}
-import uk.ac.wellcome.storage.store.{HybridIndexedStoreEntry, HybridStoreWithMaxima}
+import uk.ac.wellcome.storage.Version
+import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 import uk.ac.wellcome.storage.store.s3.S3TypedStore
+import uk.ac.wellcome.storage.store.{HybridIndexedStoreEntry, HybridStoreWithMaxima}
 
 class DynamoHybridStoreWithMaxima[Id, V, T, Metadata](
   prefix: S3ObjectLocationPrefix)(
