+++ storage/src/main/scala/uk/ac/wellcome/storage/store/dynamo/DynamoVersionedHybridStore.scala
@@ -1,6 +1,6 @@
 package uk.ac.wellcome.storage.store.dynamo
 
-import uk.ac.wellcome.storage.S3ObjectLocation
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.VersionedHybridStore
 
 class DynamoVersionedHybridStore[Id, V, T, Metadata](
