+++ storage/src/main/scala/uk/ac/wellcome/storage/store/s3/S3StreamStore.scala
@@ -4,14 +4,10 @@ import java.nio.charset.StandardCharsets
 
 import com.amazonaws.SdkClientException
 import com.amazonaws.services.s3.AmazonS3
-import com.amazonaws.services.s3.model.{
-  AmazonS3Exception,
-  ObjectMetadata,
-  PutObjectRequest,
-  S3Object
-}
+import com.amazonaws.services.s3.model.{AmazonS3Exception, ObjectMetadata, PutObjectRequest, S3Object}
 import com.amazonaws.services.s3.transfer.{TransferManagerBuilder, Upload}
 import uk.ac.wellcome.storage._
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.StreamStore
 import uk.ac.wellcome.storage.streaming._
