+++ storage/src/main/scala/uk/ac/wellcome/storage/store/s3/S3TypedStore.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.storage.store.s3
 
 import com.amazonaws.services.s3.AmazonS3
-import uk.ac.wellcome.storage.S3ObjectLocation
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.TypedStore
 import uk.ac.wellcome.storage.streaming.Codec
