+++ storage/src/main/scala/uk/ac/wellcome/storage/transfer/s3/S3PrefixTransfer.scala
@@ -2,8 +2,8 @@ package uk.ac.wellcome.storage.transfer.s3
 
 import com.amazonaws.services.s3.AmazonS3
 import uk.ac.wellcome.storage.listing.s3.S3ObjectLocationListing
+import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 import uk.ac.wellcome.storage.transfer.PrefixTransfer
-import uk.ac.wellcome.storage.{S3ObjectLocation, S3ObjectLocationPrefix}
 
 class S3PrefixTransfer()(
   implicit
