+++ storage/src/test/scala/uk/ac/wellcome/storage/listing/s3/S3ObjectLocationListingTest.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.storage.listing.s3
 
 import com.amazonaws.services.s3.AmazonS3
 import org.scalatest.Assertion
-import uk.ac.wellcome.storage.S3ObjectLocation
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 
 class S3ObjectLocationListingTest extends S3ListingTestCases[S3ObjectLocation] {
   override def assertResultCorrect(result: Iterable[S3ObjectLocation], entries: Seq[S3ObjectLocation]): Assertion =
