+++ storage/src/test/scala/uk/ac/wellcome/storage/listing/s3/S3ObjectSummaryListingTest.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.storage.listing.s3
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.S3ObjectSummary
 import org.scalatest.Assertion
-import uk.ac.wellcome.storage.S3ObjectLocation
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 
 class S3ObjectSummaryListingTest extends S3ListingTestCases[S3ObjectSummary] {
   override def assertResultCorrect(result: Iterable[S3ObjectSummary], entries: Seq[S3ObjectLocation]): Assertion = {
