+++ storage/src/test/scala/uk/ac/wellcome/storage/store/dynamo/DynamoVersionedHybridStoreTest.scala
@@ -6,9 +6,10 @@ import uk.ac.wellcome.storage.dynamo.DynamoConfig
 import uk.ac.wellcome.storage.fixtures.DynamoFixtures.Table
 import uk.ac.wellcome.storage.fixtures.{DynamoFixtures, S3Fixtures}
 import uk.ac.wellcome.storage.generators.{MetadataGenerators, Record, RecordGenerators}
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.s3.{S3StreamStore, S3TypedStore}
 import uk.ac.wellcome.storage.store.{HybridIndexedStoreEntry, HybridStoreEntry, VersionedStoreTestCases}
-import uk.ac.wellcome.storage.{S3ObjectLocation, StoreReadError, StoreWriteError, Version}
+import uk.ac.wellcome.storage.{StoreReadError, StoreWriteError, Version}
 
 class DynamoVersionedHybridStoreTest extends VersionedStoreTestCases[String, HybridStoreEntry[Record, Map[String, String]],
   DynamoHybridStoreWithMaxima[String, Int, Record, Map[String, String]]]
