+++ storage/src/test/scala/uk/ac/wellcome/storage/store/s3/S3TypedStoreTest.scala
@@ -6,6 +6,7 @@ import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.storage._
 import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
 import uk.ac.wellcome.storage.generators.{MetadataGenerators, Record, RecordGenerators}
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.fixtures.BucketNamespaceFixtures
 import uk.ac.wellcome.storage.store.{TypedStoreEntry, TypedStoreTestCases}
 import uk.ac.wellcome.storage.streaming.InputStreamWithLengthAndMetadata
