+++ storage/src/test/scala/uk/ac/wellcome/storage/transfer/s3/S3TransferTest.scala
@@ -1,8 +1,8 @@
 package uk.ac.wellcome.storage.transfer.s3
 
-import uk.ac.wellcome.storage.S3ObjectLocation
 import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
 import uk.ac.wellcome.storage.generators.{Record, RecordGenerators}
+import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.TypedStoreEntry
 import uk.ac.wellcome.storage.store.fixtures.BucketNamespaceFixtures
 import uk.ac.wellcome.storage.store.s3.S3TypedStore
